pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


[tool.poetry]
name = "my-ip"
version = "0.2.0"
description = ""
authors = ["lainiwa <kirrik96@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
toml = "^0.10.0"
pydantic = "^0.32.2"
click = "^7.0"
trio = "^0.12.1"
asks = "^2.3"
pyxdg = "^0.26.0"
loguru = "^0.3.2"
importlib_metadata = "^0.23.0"
[tool.poetry.dev-dependencies]
pytest = "^3.0"
pre-commit = "^1.18"
sphinx = "^2.2"
sphinx_rtd_theme = "^0.4.3"
sphinxcontrib-napoleon = "^0.7.0"
sphinxcontrib-golangdomain = "=0.2.0.dev0"  # had to manually install
sphinx-autoapi = "^1.1"  # otherwise `poetry add --dev sphinx-autoapi` would fail
pylint = "^2.4"
ipython = "^7.8"
mypy = "^0.730.0"
[tool.poetry.scripts]
mip = 'my_ip.console:cli'


[tool.isort]
# https://github.com/timothycrosley/isort/wiki/isort-Settings
line_length = 79
atomic = true
indent = '    '
multi_line_output = 3
length_sort = true


[tool.black]
# pre-commit run black --help
line-length = 79
verbose = true
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
